+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - By minumum require 8 characters for password
-
+- Do not show mark visited switch if user is not logged in
 ## [0.1.3] - 2025-02-20
 
 ### Added